[core] Add knownLayers API to BlackBox, ExtModule (#4983)
Add an API to add `knownlayers` specifications to FIRRTL `extmodule`s.
This is a mechanism to declare that an external module was compiled with
support for specific layers.  This then allows a FIRRTL comopiler compiler
to know that it should automatically enable these layers if needed.

In order to make this work, a number of other changes had to be made.

First, move `atModuleBodyEnd` from `RawModule` to `BaseModule`.  This
allows capturing of the `Builder` state when the module closes.  Here,
this is used to capture the layers that are known to the `Builder` if a
module is later converted to a `Definition`.

Second, the default layers are now added to the `Builder` when it starts
as opposed to when it ends.  This allows for the default layers to be
captured by module or external modules and then emitted correctly if
needed (in the `Definition` or as FIRRTL `knownlayers`).

Third, when an `Instance` is created, any layers that it has need to be
added to the `Builder`.  This can arise if a `Definition` is imported via
an annotation and used to create a blackbox which will define additional
`knownlayers`.  If these layers are not added to the FIRRTL circuit, then
this is illegal FIRRTL.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>