Fix construction of PseudoModules to have correct parent
This is mostly used when constructing PseudoModules when looking up
Instances within other Instances or Definitions. Previously, we would
construct the PseudoModule then manually set the _parent field. The
problem with this is that _parent is used during construction of the
object and defaults to Builder.currentModule. Instead, we properly set
the Builder state to hold the desired _parent and then construct the
PseudoModule.

This also cleaned up tech debt in Module.do_pseudo_apply which
duplicated logic from the standard Module.apply path in a way that had
already gotten out of sync and likely would result in other very subtle
bugs.