[core] Add Temporal inline layer, trait (#5018)
Add a new `Temporal` inline layers that are nested under each of the
builtin Assert, Assume, and Cover layers.  These layers exist because we
have observed situations where assertions which _should_ be legal in all
designs are unsupported by specific tools, though they are legal
properties by the SystemVerilog spec.

E.g., `s_eventually` is entirely unsupported by Verilator and has only
limited support in VCS.  The motivation for this PR is to have a location
where we can put these assertions and guard them off with tools that do
not support them.

Because this is expected to be problematic for all of the existing builtin
layers, this is added to each of Assert, Assume, and Cover.  Additionally,
as users have user-defined layers, the pattern of creating this `Temporal`
layer is factored into a trait, `HasTemporalInlineLayer` so that it can be
reused by others.

Add the temporal layers to Chisel's default layers.  This is necessary to
avoid these layers needing to be manually included by users.  It's
annoying that they're enabled, but because they are inline, it doesn't
create unnecessary files.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>