[core] Add LayerControl.Disable (#5019)
Add a new `LayerControl.Disable` type which allows for a variadic list of
layers to be disabled.  This complements the other existing controls,
`Enable(...)`, `EnableAll`, and `DisableAll`, particularly with the
addition of the new `Temporal` layers which you often want to only disable
for Verilator.  This makes it easier to disable just those layers, as
opposed to having to remember all the layers that need to be enabled (and
having your test silently not enable new layers if they are created).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>