[chiselsim] Explicit default CLI Verilator
When using the CLI Simulator options, set the default explicitly, as
opposed to relying on the default simulator in the `HasSimulator` class.
This will avoid possible mistakes where this could be changed outside of
the mechanisms that are visible to the CLI Simulator.  This will also just
be more clear.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>