[chiselsim] Allow CLI default settings updates
Modify the way that `CliOption`s are specified to include the ability to
set _default_ behavior when that option is _not_ specified.  This fixes a
hole in this API where there was no way to modify the settings unless an
option was specified on the command line.  This was particularly
problematic for thiings like the `-Dsimulator` argument which would
default to Verilator.  Due to the addition of temporal layers, we want to
disable these for Verilator.  However, we can only do that if the user
specified a `-Dsimulator=verilator` argument and not in the unspecified
case.  This provides an API to make this change to the `Cli.Simulator`
possible.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>