[chiselsim] Drop Temporal layers for Verilator Cli
When running Verilator via the provided `Cli.Simulator` trait, disbale all
`Temporal` layers that may exist.  This is done because these layers are
supposed to guard against temporal properties which are not supported in
Verilator.

This default-disabled behavior is intentionally only changes for the
`Cli.Simulator` as users should use more the standard ChiselSim settings
to disable the layers they want to disable.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>