[chiselsim] Drop Temporal layers for Verilator Cli (#5024)
When using the CLI Simulator options, set the default explicitly, as
opposed to relying on the default simulator in the `HasSimulator` class.
This will avoid possible mistakes where this could be changed outside of
the mechanisms that are visible to the CLI Simulator.  This will also just
be more clear.

Modify the way that `CliOption`s are specified to include the ability to
set _default_ behavior when that option is _not_ specified.  This fixes a
hole in this API where there was no way to modify the settings unless an
option was specified on the command line.  This was particularly
problematic for thiings like the `-Dsimulator` argument which would
default to Verilator.  Due to the addition of temporal layers, we want to
disable these for Verilator.  However, we can only do that if the user
specified a `-Dsimulator=verilator` argument and not in the unspecified
case.  This provides an API to make this change to the `Cli.Simulator`
possible.

When running Verilator via the provided `Cli.Simulator` trait, disbale all
`Temporal` layers that may exist.  This is done because these layers are
supposed to guard against temporal properties which are not supported in
Verilator.

This default-disabled behavior is intentionally only changes for the
`Cli.Simulator` as users should use more the standard ChiselSim settings
to disable the layers they want to disable.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>