[svsim] Add -j, --build-jobs, and --verilate-jobs (#5051)
Add Verilator compilation-time options to the Verilator svsim backend.
This is added to workaround an issue observed internally where Verilator
can crash when running with full parallelism, i.e., the default `-j 0`
that was previously hard-coded in the Verilator backend.

The default behavior is preserved, but users can now set `-j` OR
`--build-jobs`/`--verilate-jobs` individually.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>