[core] Add initial domain support (#5041)
Extend the allowed binary incompatibility for more problems in
`chisel3.internal.firrtl.ir*` and the same set of problems in
`firrtl.ir*`.  This is in support of allowing breakages that are necessary
to represent domain associations.

Add support for creating domains and domain types.  Add a built-in
ClockDomain type.

This is very basic support initially and many things will not work, namely
connections or compilation of these to anything meaningful.

Add a new method for creating a `domain_define` FIRRTL operation.  This is
used to forward a domain from one data to another, similar to
`probe.define`.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>