[core] Deprecate (experimental) ExtModule, Parms, utility traits (#5102)
Deprecate all `Param`s and `ExtModule` inside experimental as this is
moving into chisel3 proper in Chisel 8.  This is insanely difficult to do
with `case class`es, but I think we can just deprecate these as so and the
result is that all `case class` methods are deprecated and we can just
make all these `case class`es normal `class`es in Chisel 8.

Deprecate all the external module utility traits used to add methods for
including the source of the blackbox via annotations.  There's no point in
having these traits and not just putting the methods on `ExtModule`
directly.  Do this and deprecate the traits.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>