Support absolute path source locators (#5111)
The default behavior of Scala-cli is to use a Bloop server running in
the root directory for compilation. This results in absolute paths in
for source locators at compile time, but our logic had assumed we always
had relative paths and was manually stripping the leading forward slash
in source locator paths. Instead, we strip the leading forward slash for
relative paths by including it in the path prefix we remove, and then in
error reporting we add support for paths that are already absolute.