[svsim] Improve error message on illegal poke
Fix a bad error message in svsim that would occur if a user tries to poke
an output port.  Test that a set is only being done on a "settable" port
and that a get is only being done on a "gettable" port.  Note: all ports
are "gettable", so I am unable to test the latter require.

Fixes #5123.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>