Add runpath option to UnitTests main
Add a `-R` option to the `chisel3.UnitTests` main class which mirrors
the same option in scalatest's runner. This option allows the user to
supply a reduced classpath to be searched for Chisel unit test classes.
This allows discovery to be much faster in larger projects with many
dependencies, since the classpath contains all dependencies, but the
runpath can be limited to discover only test of the current project.